backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (Business logic only)
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# KERI integration - HTTP client for KERIA API
httpx==0.25.2
aiohttp==3.9.1

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
PyNaCl==1.5.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1
pydantic-extra-types==2.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Monitoring and logging
structlog==23.2.0

# Date and time
python-dateutil==2.8.2
pytz==2023.3